Bump up Java version to 11 and use junit tag to group IT and slow IT
