+++ docs/en/guides/How-to-build.md
@@ -16,7 +16,7 @@ If you need to execute build behind the proxy, edit the *.mvn/jvm.config* and se
 ```
 
 ### Building from GitHub
-1. Prepare git, JDK8+, and Maven 3.6+.
+1. Prepare git, JDK11+, and Maven 3.6+.
 1. Clone the project.
 
     If you want to build a release from source codes, set a `tag name` by using `git clone -b [tag_name] ...` while cloning.
@@ -41,7 +41,7 @@ If you need to execute build behind the proxy, edit the *.mvn/jvm.config* and se
 
 For each official Apache release, there is a complete and independent source code tar, which includes all source codes. You could download it from [SkyWalking Apache download page](http://skywalking.apache.org/downloads/). There is no requirement related to git when compiling this. Just follow these steps.
 
-1. Prepare JDK8+ and Maven 3.6+.
+1. Prepare JDK11+ and Maven 3.6+.
 1. Run `./mvnw clean package -Dmaven.test.skip`.
 1. All packages are in `/dist`.(.tar.gz for Linux and .zip for Windows).
