+++ docs/en/setup/backend/backend-setup.md
@@ -14,7 +14,7 @@ SkyWalking's backend distribution package consists of the following parts:
 
 ## Requirements and default settings
 
-Requirement: **JDK8 to JDK17 are tested**. Other versions are not tested and may or may not work.
+Requirement: **JDK11 to JDK17 are tested**. Other versions are not tested and may or may not work.
 
 Before you begin, you should understand that the main purpose of the following quickstart is to help you obtain a basic configuration for previews/demos. Performance and long-term running are **NOT** among the purposes of the quickstart.
