+++ oap-server/server-cluster-plugin/cluster-etcd-plugin/src/test/java/org/apache/skywalking/oap/server/cluster/plugin/etcd/ClusterModuleEtcdProviderFunctionalIT.java
@@ -16,20 +16,18 @@
  *
  */
 
-package org.apache.skywalking.oap.server.cluster.plugin.consul;
+package org.apache.skywalking.oap.server.cluster.plugin.etcd;
 
-import com.google.common.base.Strings;
-import com.orbitz.consul.AgentClient;
-import com.orbitz.consul.Consul;
+import io.etcd.jetcd.Client;
 import lombok.Getter;
 import org.apache.skywalking.oap.server.core.cluster.ClusterCoordinator;
 import org.apache.skywalking.oap.server.core.cluster.ClusterNodesQuery;
-import org.apache.skywalking.oap.server.core.cluster.ClusterRegister;
 import org.apache.skywalking.oap.server.core.cluster.ClusterWatcher;
 import org.apache.skywalking.oap.server.core.cluster.RemoteInstance;
 import org.apache.skywalking.oap.server.core.remote.client.Address;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.module.ModuleProvider;
+import org.apache.skywalking.oap.server.library.module.ModuleStartException;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.telemetry.TelemetryModule;
 import org.apache.skywalking.oap.server.telemetry.api.MetricsCreator;
@@ -37,10 +35,7 @@ import org.apache.skywalking.oap.server.telemetry.none.MetricsCreatorNoop;
 import org.apache.skywalking.oap.server.telemetry.none.NoneTelemetryProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.powermock.reflect.Whitebox;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.wait.strategy.Wait;
@@ -55,33 +50,32 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.mock;
 
 @Testcontainers
-@ExtendWith(MockitoExtension.class)
-public class ITClusterModuleConsulProviderFunctionalTest {
+public class ClusterModuleEtcdProviderFunctionalIT {
 
-    private String consulAddress;
+    private String endpoint;
+    private NoneTelemetryProvider telemetryProvider;
 
     @Container
     public final GenericContainer<?> container =
-        new GenericContainer<>(DockerImageName.parse("consul:0.9"))
-            .waitingFor(Wait.forLogMessage(".*Synced node info.*", 1))
-            .withCommand("agent", "-server", "-bootstrap-expect=1", "-client=0.0.0.0")
-            .withExposedPorts(8500);
-
-    @Mock
-    private ModuleManager moduleManager;
-    @Mock
-    private NoneTelemetryProvider telemetryProvider;
+        new GenericContainer<>(DockerImageName.parse("quay.io/coreos/etcd:v3.5.0"))
+            .waitingFor(Wait.forLogMessage(".*ready to serve client requests.*", 1))
+            .withEnv(Collections.singletonMap("ALLOW_NONE_AUTHENTICATION", "yes"))
+            .withCommand(
+                "etcd",
+                "--advertise-client-urls", "http://0.0.0.0:2379",
+                "--listen-client-urls", "http://0.0.0.0:2379"
+            )
+            .withExposedPorts(2379);
 
     @BeforeEach
-    public void before() {
+    public void setup() {
+        telemetryProvider = mock(NoneTelemetryProvider.class);
         Mockito.when(telemetryProvider.getService(MetricsCreator.class))
-                .thenReturn(new MetricsCreatorNoop());
-        TelemetryModule telemetryModule = Mockito.spy(TelemetryModule.class);
-        Whitebox.setInternalState(telemetryModule, "loadedProvider", telemetryProvider);
-        Mockito.when(moduleManager.find(TelemetryModule.NAME)).thenReturn(telemetryModule);
-        consulAddress = container.getHost() + ":" + container.getMappedPort(8500);
+               .thenReturn(new MetricsCreatorNoop());
+        endpoint = "http://" + container.getHost() + ":" + container.getMappedPort(2379);
     }
 
     @Test
@@ -90,7 +84,6 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         ModuleProvider provider = createProvider(serviceName);
 
         Address selfAddress = new Address("127.0.0.1", 1000, true);
-
         RemoteInstance instance = new RemoteInstance(selfAddress);
         ClusterCoordinator coordinator = getClusterCoordinator(provider);
         ClusterMockWatcher watcher = new ClusterMockWatcher();
@@ -101,7 +94,6 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         List<RemoteInstance> remoteInstances = notifiedRemoteNodes(watcher, 1);
         assertEquals(1, remoteInstances.size());
         assertEquals(1,  queryRemoteNodes(provider, 1).size());
-
         Address queryAddress = remoteInstances.get(0).getAddress();
         assertEquals(selfAddress, queryAddress);
         assertTrue(queryAddress.isSelf());
@@ -110,9 +102,9 @@ public class ITClusterModuleConsulProviderFunctionalTest {
     @Test
     public void registerRemoteOfInternal() throws Exception {
         final String serviceName = "register_remote_internal";
-        ModuleProvider provider = createProvider(serviceName, "127.0.1.2", 1001);
+        ModuleProvider provider = createProvider(serviceName, "127.0.1.2", 1000);
 
-        Address selfAddress = new Address("127.0.0.2", 1002, true);
+        Address selfAddress = new Address("127.0.0.2", 1000, true);
         RemoteInstance instance = new RemoteInstance(selfAddress);
         ClusterCoordinator coordinator = getClusterCoordinator(provider);
         ClusterMockWatcher watcher = new ClusterMockWatcher();
@@ -126,7 +118,7 @@ public class ITClusterModuleConsulProviderFunctionalTest {
 
         Address queryAddress = remoteInstances.get(0).getAddress();
         assertEquals("127.0.1.2", queryAddress.getHost());
-        assertEquals(1001, queryAddress.getPort());
+        assertEquals(1000, queryAddress.getPort());
         assertTrue(queryAddress.isSelf());
     }
 
@@ -142,7 +134,7 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         coordinatorB.registerWatcher(watcherB);
         coordinatorB.start();
         // Mixed or Aggregator
-        Address selfAddress = new Address("127.0.0.3", 1003, true);
+        Address selfAddress = new Address("127.0.0.3", 1000, true);
         RemoteInstance instance = new RemoteInstance(selfAddress);
         coordinatorA.start();
         coordinatorA.registerRemote(instance);
@@ -170,8 +162,8 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         coordinatorB.registerWatcher(watcherB);
         coordinatorB.start();
 
-        Address addressA = new Address("127.0.0.4", 1004, true);
-        Address addressB = new Address("127.0.0.5", 1005, true);
+        Address addressA = new Address("127.0.0.4", 1000, true);
+        Address addressB = new Address("127.0.0.5", 1000, true);
 
         RemoteInstance instanceA = new RemoteInstance(addressA);
         RemoteInstance instanceB = new RemoteInstance(addressB);
@@ -201,9 +193,8 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         ClusterMockWatcher watcherB = new ClusterMockWatcher();
         coordinatorB.registerWatcher(watcherB);
         coordinatorB.start();
-
-        Address addressA = new Address("127.0.0.6", 1006, true);
-        Address addressB = new Address("127.0.0.7", 1007, true);
+        Address addressA = new Address("127.0.0.4", 1000, true);
+        Address addressB = new Address("127.0.0.5", 1000, true);
 
         RemoteInstance instanceA = new RemoteInstance(addressA);
         RemoteInstance instanceB = new RemoteInstance(addressB);
@@ -220,9 +211,8 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         assertEquals(2,  queryRemoteNodes(providerB, 2).size());
 
         // unregister A
-        Consul client = Whitebox.getInternalState(providerA, "client");
-        AgentClient agentClient = client.agentClient();
-        agentClient.deregister(instanceA.getAddress().toString());
+        Client client = Whitebox.getInternalState(coordinatorA, "client");
+        client.close();
 
         // only B
         remoteInstancesOfB = notifiedRemoteNodes(watcherB, 1, 120);
@@ -233,18 +223,20 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         assertTrue(addressB.isSelf());
     }
 
-    private ClusterModuleConsulProvider createProvider(String serviceName) throws Exception {
+    private ClusterModuleEtcdProvider createProvider(String serviceName)
+        throws ModuleStartException {
         return createProvider(serviceName, null, 0);
     }
 
-    private ClusterModuleConsulProvider createProvider(String serviceName, String internalComHost,
-        int internalComPort) throws Exception {
-        ClusterModuleConsulProvider provider = new ClusterModuleConsulProvider();
+    private ClusterModuleEtcdProvider createProvider(String serviceName, String internalComHost,
+                                                     int internalComPort)
+        throws ModuleStartException {
+        ClusterModuleEtcdProvider provider = new ClusterModuleEtcdProvider();
 
-        ClusterModuleConsulConfig config = new ClusterModuleConsulConfig();
+        ClusterModuleEtcdConfig config = new ClusterModuleEtcdConfig();
         provider.newConfigCreator().onInitialized(config);
 
-        config.setHostPort(consulAddress);
+        config.setEndpoints(endpoint);
         config.setServiceName(serviceName);
 
         if (!StringUtil.isEmpty(internalComHost)) {
@@ -254,47 +246,24 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         if (internalComPort > 0) {
             config.setInternalComPort(internalComPort);
         }
-        provider.setManager(moduleManager);
+        TelemetryModule telemetryModule = Mockito.spy(TelemetryModule.class);
+        Whitebox.setInternalState(telemetryModule, "loadedProvider", telemetryProvider);
+        ModuleManager manager = mock(ModuleManager.class);
+        Mockito.when(manager.find(TelemetryModule.NAME)).thenReturn(telemetryModule);
+
+        provider.setManager(manager);
         provider.prepare();
         provider.start();
         provider.notifyAfterCompleted();
-
         return provider;
-    }
-
-    private boolean needUsingInternalAddr(ClusterModuleConsulConfig config) {
-        return !Strings.isNullOrEmpty(config.getInternalComHost()) && config.getInternalComPort() > 0;
     }
 
     private ClusterCoordinator getClusterCoordinator(ModuleProvider provider) {
         return provider.getService(ClusterCoordinator.class);
     }
 
-    private ClusterRegister getClusterRegister(ModuleProvider provider) {
-        return provider.getService(ClusterRegister.class);
-    }
-
     private ClusterNodesQuery getClusterNodesQuery(ModuleProvider provider) {
         return provider.getService(ClusterNodesQuery.class);
-    }
-
-    private List<RemoteInstance> notifiedRemoteNodes(ClusterMockWatcher watcher, int goals)
-        throws InterruptedException {
-        return notifiedRemoteNodes(watcher, goals, 20);
-    }
-
-    private List<RemoteInstance> notifiedRemoteNodes(ClusterMockWatcher watcher, int goals,
-                                                     int cyclic) throws InterruptedException {
-        do {
-            List<RemoteInstance> instances = watcher.getRemoteInstances();
-            if (instances.size() == goals) {
-                return instances;
-            } else {
-                Thread.sleep(1000);
-            }
-        }
-        while (--cyclic > 0);
-        return Collections.emptyList();
     }
 
     private List<RemoteInstance> queryRemoteNodes(ModuleProvider provider, int goals) throws InterruptedException {
@@ -302,7 +271,7 @@ public class ITClusterModuleConsulProviderFunctionalTest {
     }
 
     private List<RemoteInstance> queryRemoteNodes(ModuleProvider provider, int goals,
-        int cyclic) throws InterruptedException {
+                                                  int cyclic) throws InterruptedException {
         do {
             List<RemoteInstance> instances = getClusterNodesQuery(provider).queryRemoteNodes();
             if (instances.size() == goals) {
@@ -315,7 +284,27 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         return Collections.emptyList();
     }
 
-    private void validateServiceInstance(Address selfAddress, Address otherAddress, List<RemoteInstance> queryResult) {
+    private List<RemoteInstance> notifiedRemoteNodes(ClusterMockWatcher watcher, int goals)
+        throws InterruptedException {
+        return notifiedRemoteNodes(watcher, goals, 20);
+    }
+
+    private List<RemoteInstance> notifiedRemoteNodes(ClusterMockWatcher watcher, int goals,
+                                                  int cyclic) throws InterruptedException {
+        do {
+            List<RemoteInstance> instances = watcher.getRemoteInstances();
+            if (instances.size() == goals) {
+                return instances;
+            } else {
+                Thread.sleep(1000);
+            }
+        }
+        while (--cyclic > 0);
+        return Collections.emptyList();
+    }
+
+    private void validateServiceInstance(Address selfAddress, Address otherAddress,
+                                         List<RemoteInstance> queryResult) {
         assertEquals(2, queryResult.size());
 
         boolean selfExist = false, otherExist = false;
@@ -333,7 +322,7 @@ public class ITClusterModuleConsulProviderFunctionalTest {
         assertTrue(otherExist);
     }
 
-    class ClusterMockWatcher implements ClusterWatcher {
+    static class ClusterMockWatcher implements ClusterWatcher {
         @Getter
         private List<RemoteInstance> remoteInstances = new ArrayList<>();
