+++ oap-server/server-configuration/configuration-etcd/src/test/java/org/apache/skywalking/oap/server/configuration/etcd/EtcdConfigurationIT.java
@@ -51,7 +51,7 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 
 @Slf4j
 @Testcontainers
-public class ITEtcdConfigurationTest {
+public class EtcdConfigurationIT {
     @Container
     public final GenericContainer<?> container =
         new GenericContainer<>(DockerImageName.parse("quay.io/coreos/etcd:v3.5.0"))
