+++ oap-server/server-configuration/configuration-zookeeper/src/test/java/org/apache/skywalking/oap/server/configuration/zookeeper/it/ZookeeperConfigurationIT.java
@@ -16,12 +16,13 @@
  *
  */
 
-package org.apache.skywalking.oap.server.configuration.nacos;
+package org.apache.skywalking.oap.server.configuration.zookeeper.it;
 
-import com.alibaba.nacos.api.NacosFactory;
-import com.alibaba.nacos.api.config.ConfigService;
-import com.alibaba.nacos.api.exception.NacosException;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.curator.RetryPolicy;
+import org.apache.curator.framework.CuratorFramework;
+import org.apache.curator.framework.CuratorFrameworkFactory;
+import org.apache.curator.retry.ExponentialBackoffRetry;
 import org.apache.skywalking.oap.server.library.module.ApplicationConfiguration;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
@@ -39,7 +40,6 @@ import org.yaml.snakeyaml.Yaml;
 
 import java.io.FileNotFoundException;
 import java.io.Reader;
-import java.util.Collections;
 import java.util.Map;
 import java.util.Properties;
 
@@ -50,22 +50,23 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @Slf4j
 @Testcontainers
-public class ITNacosConfigurationTest {
+public class ZookeeperConfigurationIT {
     private final Yaml yaml = new Yaml();
 
-    private NacosConfigurationTestProvider provider;
+    private MockZookeeperConfigurationProvider provider;
 
     @Container
     public final GenericContainer<?> container =
-        new GenericContainer<>(DockerImageName.parse("nacos/nacos-server:1.4.2"))
-            .waitingFor(Wait.forLogMessage(".*Nacos started successfully.*", 1))
-            .withEnv(Collections.singletonMap("MODE", "standalone"))
-            .withExposedPorts(8848);
+        new GenericContainer<>(DockerImageName.parse("zookeeper:3.5"))
+            .waitingFor(Wait.forLogMessage(".*binding to port.*", 1))
+            .withExposedPorts(2181);
+
+    private String zkAddress;
 
     @BeforeEach
     public void setUp() throws Exception {
-        System.setProperty("nacos.host", container.getHost());
-        System.setProperty("nacos.port", String.valueOf(container.getMappedPort(8848)));
+        zkAddress = container.getHost() + ":" + container.getMappedPort(2181);
+        System.setProperty("zk.address", zkAddress);
 
         final ApplicationConfiguration applicationConfiguration = new ApplicationConfiguration();
         loadConfig(applicationConfiguration);
@@ -73,7 +74,8 @@ public class ITNacosConfigurationTest {
         final ModuleManager moduleManager = new ModuleManager();
         moduleManager.init(applicationConfiguration);
 
-        provider = (NacosConfigurationTestProvider) moduleManager.find(NacosConfigurationTestModule.NAME).provider();
+        provider = (MockZookeeperConfigurationProvider) moduleManager.find(MockZookeeperConfigurationModule.NAME)
+                                                                     .provider();
 
         assertNotNull(provider);
     }
@@ -81,24 +83,24 @@ public class ITNacosConfigurationTest {
     @SuppressWarnings("StatementWithEmptyBody")
     @Test
     @Timeout(20)
-    public void shouldReadUpdated() throws NacosException {
+    public void shouldReadUpdated() throws Exception {
+        String namespace = "/default";
+        String key = "test-module.default.testKey";
         assertNull(provider.watcher.value());
 
-        final Properties properties = new Properties();
-        final String nacosHost = System.getProperty("nacos.host");
-        final String nacosPort = System.getProperty("nacos.port");
-        log.info("nacosHost: {}, nacosPort: {}", nacosHost, nacosPort);
-        properties.put("serverAddr", nacosHost + ":" + nacosPort);
+        RetryPolicy retryPolicy = new ExponentialBackoffRetry(1000, 3);
+        CuratorFramework client = CuratorFrameworkFactory.newClient(zkAddress, retryPolicy);
+        client.start();
+        log.info("per path: " + namespace + "/" + key);
 
-        final ConfigService configService = NacosFactory.createConfigService(properties);
-        assertTrue(configService.publishConfig("test-module.default.testKey", "skywalking", "500"));
+        assertTrue(client.create().creatingParentsIfNeeded().forPath(namespace + "/" + key, "500".getBytes()) != null);
+
+        log.info("data: " + new String(client.getData().forPath(namespace + "/" + key)));
 
         for (String v = provider.watcher.value(); v == null; v = provider.watcher.value()) {
         }
 
-        assertEquals("500", provider.watcher.value());
-
-        assertTrue(configService.removeConfig("test-module.default.testKey", "skywalking"));
+        assertTrue(client.delete().forPath(namespace + "/" + key) == null);
 
         for (String v = provider.watcher.value(); v != null; v = provider.watcher.value()) {
         }
@@ -108,47 +110,37 @@ public class ITNacosConfigurationTest {
 
     @Test
     @Timeout(20)
-    public void shouldReadUpdatedGroup() throws NacosException {
+    public void shouldReadUpdated4GroupConfig() throws Exception {
+        String namespace = "/default";
+        String key = "test-module.default.testKeyGroup";
         assertEquals("{}", provider.groupWatcher.groupItems().toString());
 
-        final Properties properties = new Properties();
-        final String nacosHost = System.getProperty("nacos.host");
-        final String nacosPort = System.getProperty("nacos.port");
-        log.info("nacosHost: {}, nacosPort: {}", nacosHost, nacosPort);
-        properties.put("serverAddr", nacosHost + ":" + nacosPort);
+        RetryPolicy retryPolicy = new ExponentialBackoffRetry(1000, 3);
+        CuratorFramework client = CuratorFrameworkFactory.newClient(zkAddress, retryPolicy);
+        client.start();
+        log.info("per path: " + namespace + "/" + key);
 
-        final ConfigService configService = NacosFactory.createConfigService(properties);
-        //test add group key and item1 item2
-        assertTrue(configService.publishConfig("test-module.default.testKeyGroup", "skywalking", "item1\n item2"));
-        assertTrue(configService.publishConfig("item1", "skywalking", "100"));
-        assertTrue(configService.publishConfig("item2", "skywalking", "200"));
+        assertTrue(client.create().creatingParentsIfNeeded().forPath(namespace + "/" + key + "/item1", "100".getBytes()) != null);
+        assertTrue(client.create().creatingParentsIfNeeded().forPath(namespace + "/" + key + "/item2", "200".getBytes()) != null);
+
+        log.info("data: " + new String(client.getData().forPath(namespace + "/" + key + "/item1")));
+        log.info("data: " + new String(client.getData().forPath(namespace + "/" + key + "/item2")));
+
         for (String v = provider.groupWatcher.groupItems().get("item1"); v == null; v = provider.groupWatcher.groupItems().get("item1")) {
         }
         for (String v = provider.groupWatcher.groupItems().get("item2"); v == null; v = provider.groupWatcher.groupItems().get("item2")) {
         }
-        assertEquals("100", provider.groupWatcher.groupItems().get("item1"));
-        assertEquals("200", provider.groupWatcher.groupItems().get("item2"));
 
-        //test remove item1
-        assertTrue(configService.removeConfig("item1", "skywalking"));
+        assertTrue(client.delete().forPath(namespace + "/" + key + "/item1") == null);
+        assertTrue(client.delete().forPath(namespace + "/" + key + "/item2") == null);
+
         for (String v = provider.groupWatcher.groupItems().get("item1"); v != null; v = provider.groupWatcher.groupItems().get("item1")) {
         }
-        assertNull(provider.groupWatcher.groupItems().get("item1"));
-
-        //test modify item1
-        assertTrue(configService.publishConfig("item1", "skywalking", "300"));
-        for (String v = provider.groupWatcher.groupItems().get("item1"); v == null; v = provider.groupWatcher.groupItems().get("item1")) {
-        }
-        assertEquals("300", provider.groupWatcher.groupItems().get("item1"));
-
-        //test remove group key
-        assertTrue(configService.removeConfig("test-module.default.testKeyGroup", "skywalking"));
         for (String v = provider.groupWatcher.groupItems().get("item2"); v != null; v = provider.groupWatcher.groupItems().get("item2")) {
         }
+
+        assertNull(provider.groupWatcher.groupItems().get("item1"));
         assertNull(provider.groupWatcher.groupItems().get("item2"));
-        //chean
-        assertTrue(configService.removeConfig("item1", "skywalking"));
-        assertTrue(configService.removeConfig("item2", "skywalking"));
     }
 
     @SuppressWarnings("unchecked")
@@ -173,7 +165,7 @@ public class ITNacosConfigurationTest {
                         moduleConfiguration.addProviderConfiguration(name, properties);
                     });
                 }
-            });
+             });
         }
     }
 }
