+++ oap-server/server-library/library-client/src/test/java/org/apache/skywalking/library/elasticsearch/bulk/ElasticSearchIT.java
@@ -44,7 +44,7 @@ import java.util.Optional;
 import java.util.function.Function;
 
 @Slf4j
-public class ITElasticSearch {
+public class ElasticSearchIT {
 
     public static Collection<Object[]> versions() {
         // noinspection resource
