+++ oap-server/server-library/library-util/src/main/java/org/apache/skywalking/oap/server/library/util/StringUtil.java
@@ -18,8 +18,6 @@
 
 package org.apache.skywalking.oap.server.library.util;
 
-import java.util.function.Consumer;
-
 public final class StringUtil {
     public static boolean isEmpty(String str) {
         return str == null || str.length() == 0;
@@ -35,12 +33,6 @@ public final class StringUtil {
 
     public static boolean isNotBlank(String str) {
         return !isBlank(str);
-    }
-
-    public static void setIfPresent(String value, Consumer<String> setter) {
-        if (isNotEmpty(value)) {
-            setter.accept(value);
-        }
     }
 
     public static String join(final char delimiter, final String... strings) {
